Explicitly set tdb internal flash size for PSOC6
The default computation assumes that a flash sector is several times
larger than a flash page. On PSoC 6 targets this is not the case
(the two values are the same) so the computed size is too small.
This is a similar change to 5a4e370075b59089dd583de9ec0f92acc75f18fb,
but for devices which implement TDB in internal storage.