Remove unsupported folder
Mbed OS was carrying this directory for a long time. The main reason was
to enable users to benefit from some form of USB stack. Now when there
is proper implementation of USB stack in Mbed I think it's time for
unsupported directory to go. If any other functionality present in this
directory will be needed it'll need to be introduced in Mbed OS
properly.