Add Timeout rescheduling test
The `Timeout` drift test uses rescheduling inside a callback, but it is
currently disabled due to lack of stability. Rescheduling using relative
timeouts inside the callback is a bad technique as it leads to drift, so
I understand the test being disabled. It is better to use `Ticker` for a
periodic callback or to use `Timeout::attach_absolute`.

Add a simpler test that just ensures the callback is called repeatedly -
this test would detect issue #12940, and should not have stability
problems.