LWIPStack: set sockets non-blocking
When sockets were opened, they were not set non-blocking, but only had
their timeout set to 1ms.

TCP sockets were set to non-blocking on connect or accept, but UDP
sockets were left blocking with 1ms timeout.

This led to unnecessary delays when checking for data from an empty UDP
socket.

Change sockets to be non-blocking on open instead of setting the
timeout. As TCP sockets were already non-blocking, the necessary event
handling is already in place, and no other changes are required.